🎨 rename packages to _server for backward comp
