+++ caikit/runtime/grpc_server/servicers/info_servicer.py
