+++ tests/runtime/grpc_server/servicers/test_global_train_servicer_impl.py
@@ -29,12 +29,14 @@ from caikit.config import get_config
 from caikit.core import MODEL_MANAGER
 from caikit.core.data_model.producer import ProducerId
 from caikit.interfaces.common.data_model.stream_sources import S3Path
-from caikit.runtime.grpc.service_factory import (
+from caikit.runtime.grpc_server.service_factory import (
     get_inference_request,
     get_train_params,
     get_train_request,
 )
-from caikit.runtime.grpc.servicers.global_train_servicer import GlobalTrainServicer
+from caikit.runtime.grpc_server.servicers.global_train_servicer import (
+    GlobalTrainServicer,
+)
 from caikit.runtime.types.caikit_runtime_exception import CaikitRuntimeException
 from sample_lib.data_model.sample import (
     OtherOutputType,
@@ -47,7 +49,7 @@ from sample_lib.data_model.sample import (
 from sample_lib.modules import CompositeModule, OtherModule, SampleModule
 from tests.conftest import random_test_id, set_use_subprocess
 from tests.fixtures import Fixtures
-from tests.runtime.grpc.conftest import register_trained_model
+from tests.runtime.grpc_server.conftest import register_trained_model
 import caikit.core
 
 ## Helpers #####################################################################
