+++ tests/runtime/grpc_server/servicers/test_model_runtime_servicer_impl.py
@@ -27,7 +27,7 @@ import unittest
 
 # Local
 from caikit import get_config
-from caikit.runtime.grpc.servicers.model_runtime_servicer import (
+from caikit.runtime.grpc_server.servicers.model_runtime_servicer import (
     ModelRuntimeServicerImpl,
 )
 from caikit.runtime.protobufs import model_runtime_pb2
