+++ tests/runtime/grpc_server/servicers/test_model_train_servicer_impl.py
@@ -26,7 +26,9 @@ import pytest
 # Local
 from caikit.core import MODEL_MANAGER
 from caikit.core.data_model import TrainingStatus
-from caikit.runtime.grpc.servicers.model_train_servicer import ModelTrainServicerImpl
+from caikit.runtime.grpc_server.servicers.model_train_servicer import (
+    ModelTrainServicerImpl,
+)
 from caikit.runtime.protobufs import process_pb2
 from caikit.runtime.types.caikit_runtime_exception import CaikitRuntimeException
 from tests.conftest import set_use_subprocess, temp_config
