+++ tests/runtime/http_server/test_utils.py
@@ -18,8 +18,8 @@ from typing import List, Optional, Union
 # Local
 from caikit.core import DataObjectBase, dataobject
 from caikit.interfaces.common.data_model import File
-from caikit.runtime.http.pydantic_wrapper import dataobject_to_pydantic
-from caikit.runtime.http.utils import (
+from caikit.runtime.http_server.pydantic_wrapper import dataobject_to_pydantic
+from caikit.runtime.http_server.utils import (
     convert_json_schema_to_multipart,
     flatten_json_schema,
     update_dict_at_dot_path,
